#944 extract v-app-bar to component
